Backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==5.2.12
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
firebase-admin==6.6.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Pillow==12.1.1
psycopg==3.3.2
PyJWT==2.11.0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
